+++ Loader/JsonConfigurationLoader.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Loader;
 
 use EXSyst\Component\Swagger\Swagger;
+use Linkin\Bundle\SwaggerResolverBundle\Merger\OperationParameterMerger;
 
 /**
  * @author Viktor Linkin <adrenalinkin@gmail.com>
@@ -26,11 +27,12 @@ class JsonConfigurationLoader extends AbstractFileConfigurationLoader
     private $pathToFile;
 
     /**
+     * @param OperationParameterMerger $parameterMerger
      * @param string $pathToFile
      */
-    public function __construct(string $pathToFile)
+    public function __construct(OperationParameterMerger $parameterMerger, string $pathToFile)
     {
-        parent::__construct($pathToFile);
+        parent::__construct($parameterMerger, $pathToFile);
 
         $this->pathToFile = $pathToFile;
     }
@@ -38,7 +40,7 @@ class JsonConfigurationLoader extends AbstractFileConfigurationLoader
     /**
      * {@inheritdoc}
      */
-    public function loadConfiguration(): Swagger
+    protected function loadConfiguration(): Swagger
     {
         return Swagger::fromFile($this->pathToFile);
     }