+++ Loader/SwaggerPhpConfigurationLoader.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Loader;
 
 use EXSyst\Component\Swagger\Swagger;
+use Linkin\Bundle\SwaggerResolverBundle\Merger\OperationParameterMerger;
 use Symfony\Component\Routing\RouterInterface;
 use function json_decode;
 use function Swagger\scan;
@@ -34,13 +35,14 @@ class SwaggerPhpConfigurationLoader extends AbstractAnnotationConfigurationLoade
     private $scan;
 
     /**
+     * @param OperationParameterMerger $merger
      * @param RouterInterface $router
      * @param array $scan
      * @param array $exclude
      */
-    public function __construct(RouterInterface $router, array $scan, array $exclude)
+    public function __construct(OperationParameterMerger $merger, RouterInterface $router, array $scan, array $exclude)
     {
-        parent::__construct($router);
+        parent::__construct($merger, $router);
 
         $this->scan = $scan;
         $this->exclude = $exclude;
@@ -49,15 +51,13 @@ class SwaggerPhpConfigurationLoader extends AbstractAnnotationConfigurationLoade
     /**
      * {@inheritdoc}
      */
-    public function loadConfiguration(): Swagger
+    protected function loadConfiguration(): Swagger
     {
         $swaggerAnnotation = scan($this->scan, [
             'exclude' => $this->exclude,
         ]);
 
         $swagger = new Swagger(json_decode((string) $swaggerAnnotation, true));
-
-        $this->registerResources($swagger);
 
         return $swagger;
     }