+++ Loader/YamlConfigurationLoader.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Loader;
 
 use EXSyst\Component\Swagger\Swagger;
+use Linkin\Bundle\SwaggerResolverBundle\Merger\OperationParameterMerger;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -27,11 +28,12 @@ class YamlConfigurationLoader extends AbstractFileConfigurationLoader
     private $pathToFile;
 
     /**
+     * @param OperationParameterMerger $parameterMerger
      * @param string $pathToFile
      */
-    public function __construct(string $pathToFile)
+    public function __construct(OperationParameterMerger $parameterMerger, string $pathToFile)
     {
-        parent::__construct($pathToFile);
+        parent::__construct($parameterMerger, $pathToFile);
 
         $this->pathToFile = $pathToFile;
     }
@@ -39,7 +41,7 @@ class YamlConfigurationLoader extends AbstractFileConfigurationLoader
     /**
      * {@inheritdoc}
      */
-    public function loadConfiguration(): Swagger
+    protected function loadConfiguration(): Swagger
     {
         return new Swagger(Yaml::parseFile($this->pathToFile));
     }