+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 ## [Unreleased]
 ### Added
 - Added symfony cache warmer for the swagger configuration and enable him by default.
+- Added console notification when some api definitions have not reference to the source file.
 - Added automatic cache warm up in the debug mode, according to source file modification.
 - Added composer requirement: `symfony/yaml`.
 ### Changed
 - Extend `NormalizationFailedException` from `InvalidOptionsException` instead `RuntimeException`.
 - Removed possibility set `MergeStrategyInterface` for single call `SwaggerResolverFactory::createForRequest`.
 - Renamed `PathParameterMerger` into `OperationParameterMerger`.
+- Reworked `SwaggerConfigurationLoaderInterface`.
 
 ## [0.3.0] - 2019-03-03
 ### Added